In progress PR 315 fixes
