+++ lib/bitcoin/IBitcoinConfig.ts
@@ -1,4 +1,4 @@
-import IProofOfFeeConfig from './IProofOfFeeConfig';
+import ProofOfFeeConfig from './pof/models/ProofOfFeeConfig';
 
 /**
  * Defines all the configuration parameters needed to initialize Sidetree bitcoin service.
@@ -18,5 +18,5 @@ export interface IBitcoinConfig {
   requestTimeoutInMilliseconds: number | undefined;
   requestMaxRetries: number | undefined;
   transactionPollPeriodInSeconds: number | undefined;
-  proofOfFeeConfig: IProofOfFeeConfig;
+  proofOfFeeConfig: ProofOfFeeConfig;
 }
