+++ lib/bitcoin/pof/models/QuantileInfo.ts
@@ -1,0 +1,9 @@
+/**
+ * Quantile information stored in the quantile store
+ * for each group.
+ */
+export default interface QuantileInfo {
+  groupId: number;
+  quantile: number;
+  groupFreqVector: number[];
+}
